+++ src/atoms/member.js
@@ -4,3 +4,18 @@ export const isLoggedUser = atom({
   key: 'isLoggedUser',
   default: false,
 });
+
+export const nickNameState = atom({
+  key: 'nickname',
+  default: '',
+});
+
+export const tokenState = atom({
+  key: 'token',
+  default: '',
+});
+
+export const uuidState = atom({
+  key: 'uuid',
+  default: '',
+});
