--- src/atoms/nickname.js
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const nicknameState = atom({
-  key: 'nickname',
-  default: '',
-});
