--- src/atoms/token.js
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const tokenState = atom({
-  key: 'token',
-  default: '',
-});
