--- src/atoms/uuid.js
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const uuidState = atom({
-  key: 'uuid',
-  default: '',
-});
